Create README.md
The attempt is to isolate and measure the natural frequencies of the mind and body. Singling out the htghest level of natural and reactive states that exposure is meant to activate the healing and calming process. Introducing audible sounds through a full spectrum of vibrational applications within the most extensive and measurable range that proves consistant, systematic data. Then compare that to others data and see if it holds water.